+++ native/src/Main.cpp
@@ -564,7 +564,7 @@ RED4EXT_C_EXPORT void RED4EXT_CALL Query(RED4ext::v1::PluginInfo* info)
 {
     info->name = L"Character Preset Manager ACU Import Service";
     info->author = L"dklyntly";
-    info->version = RED4EXT_V1_SEMVER(3, 0, 10);
+    info->version = RED4EXT_V1_SEMVER(3, 1, 0);
     info->runtime = RED4EXT_V1_RUNTIME_VERSION_INDEPENDENT;
     info->sdk = RED4EXT_V1_SDK_VERSION_CURRENT;
 }